requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Templates
jinja2==3.1.4

# Database
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.10

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
itsdangerous==2.2.0

# Forms and Validation
pydantic==2.9.2
pydantic-settings==2.6.0
email-validator==2.2.0

# Image Processing
pillow==11.0.0

# Search
whoosh==2.7.4

# Utilities
python-slugify==8.0.4
httpx==0.27.2

# Development
pytest==8.3.3
pytest-asyncio==0.24.0
black==24.10.0
ruff==0.7.4
